pyproject.toml
[project]
name = "youtube-pipeline"
version = "0.1.0"
description = "YouTube Video Automation Pipeline"
requires-python = ">=3.12"
dependencies = [
    "temporalio>=1.24.0",
    "fastapi>=0.135.0",
    "uvicorn>=0.42.0",
    "sqlmodel>=0.0.37",
    "alembic>=1.18.0",
    "pydantic-settings>=2.13.0",
    "gspread>=6.2.0",
    "google-auth",
    "structlog>=25.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=9.0.0",
    "pytest-asyncio>=1.3.0",
    "pytest-cov>=7.0.0",
    "ruff",
    "mypy",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"

[tool.ruff]
target-version = "py312"
line-length = 88

[tool.ruff.lint]
select = ["E", "F", "I"]

[tool.mypy]
python_version = "3.12"
strict = true

[dependency-groups]
dev = [
    "mypy>=1.20.0",
]
